requirements.txt
Flask~=3.1.2
requests~=2.32.5
openai~=2.8.1
newspaper3k~=0.2.8
python-dotenv~=1.2.1
playwright